requirements.txt
Flask==2.0.2
numpy==1.21.2
pandas==1.4
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.35
gunicorn==20.1.0
botocore==1.25.0
boto3==1.22.0
pillow==9.1.1
pathlib==1.0.1
flask-cors==3.0.10